refactor: useGetProductDetail api 로직 axios, path(constant), hook 으로 추상화
