+++ src/api/hooks/useGetProductDetail.ts
@@ -1,48 +1,11 @@
 import { useQuery } from '@tanstack/react-query';
 
-import type { GoodsData } from '@/types';
-
-import { fetchInstance } from '../instance';
-
-interface ProductDetaiResponseData {
-  detail: GoodsData & {
-    isAccessableProductPage: boolean;
-    review: {
-      averageRating: number;
-      totalReviewCount: number;
-    };
-    productDescription: {
-      images: string[];
-    };
-    productDetailInfo: {
-      announcements: {
-        displayOrder: number;
-        name: string;
-        value: string;
-      }[];
-      terms: {
-        displayOrder: number;
-        title: string;
-        description: string;
-      }[];
-    };
-  };
-}
-
-const getProductDetailPath = (productId: string) => `/v1/products/${productId}/detail`;
-const productDetailQueryKey = (productId: string) => [getProductDetailPath(productId)];
-
-export const getProductDetail = async (productId: string) => {
-  const response = await fetchInstance.get<ProductDetaiResponseData>(
-    getProductDetailPath(productId),
-  );
-
-  return response.data;
-};
+import { getProductDetail } from '../apis/product';
+import API from '../path';
 
 export const useGetProductDetail = (productId: string) => {
   return useQuery({
-    queryKey: productDetailQueryKey(productId),
+    queryKey: [API.PRODUCT.DETAIL(productId)],
     queryFn: async () => {
       const res = await getProductDetail(productId);
       return res;
